# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple microservice for borrowing books with jwt auth",
        "title": "Book microservice",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Dias Galikhanov",
            "email": "diasgalikhanov@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8586",
    "basePath": "/api/v1",
    "paths": {
        "/book": {
            "get": {
                "description": "list books",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "List Books",
                "operationId": "list-books",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Book"
                            }
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Create Book",
                "operationId": "create-book",
                "parameters": [
                    {
                        "description": "book info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/book/{book_id}": {
            "get": {
                "description": "get book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Get Book",
                "operationId": "get-book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "book id",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Update Book",
                "operationId": "update-book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "book id",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "book info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Delete Book",
                "operationId": "delete-book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "book id",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}